pyproject.toml
[tool.poetry]
name = "openmethane_prior"
version = "0.0.1"
description = "Method to calculate a gridded, prior emissions estimate for methane across Australia."
authors = [
     "Peter Rayner <peter.rayner@superpowerinstitute.com.au>",
     "Gerard Mason <gerard.mason@superpowerinstitute.com.au>",
     "Jared Lewis <jared.lewis@climate-resource.com>",
     "Daniel Busch <daniel.busch@climate-resource.com>",
]
readme = "README.md"
packages = [
    {include = "openmethane_prior", from = "src"},
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
numpy = "^1.26.4"
netcdf4 = "^1.6.5"
xarray = "^2024.5.0"
rioxarray = "^0.15.5"
pyproj = "^3.6.1"
pandas = "^2.2.2"
geopandas = "^0.14.4"
python-dotenv = "^1.0.1"
colorama = "^0.4.6"
segment-geospatial = "^0.10.6"
cdsapi = "^0.7.0"
shapely = "^2.0.4"
environs = "^11.0.0"

[tool.poetry.group.tests.dependencies]
pytest = "^8.2.1"
pytest-regressions = "^2.5.0"

[tool.poetry.group.dev.dependencies]
licenseheaders = "^0.8.8"
ruff = "^0.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
src = ["src"]
target-version = "py311"
# Provide some leeway for long docstring
line-length = 100


[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    "E",  # pycodestyle error
    "W",  # pycodestyle warning
    "F",  # pyflakes
    "I",  # isort
    "D",  # pydocstyle
    "PL",  # pylint
    "TRY",  # tryceratops
    "NPY",  # numpy rules
    "RUF",  # ruff specifics
    "UP",  # pyupgrade
    "S",  # flake8-bandit
    # pandas support via pandas-vet. In some cases we will want to disable
    # this because it can lead to too many false positives.
    "PD",
]
unfixable = [
    "PD002",  # Disable autofix for inplace as this often introduces bugs
]
ignore = [
    "D100",  # Missing docstring in public module
    "D105",  # Missing docstring in magic methods
    "D106",  # Missing docstring in public nested class
    "D400",  # First line should end with a period
    "TRY003",  # Avoid specifying long messages outside the exception class
    "PD011", #  Use `.to_numpy()` instead of `.values`
    "S101",  # Use of `assert` detected
]

[tool.ruff.lint.per-file-ignores]
"test*.py" = [
    "D",  # Documentation not needed in tests
    "S101",  # S101 Use of `assert` detected
    "PLR2004" # Magic value used in comparison
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.isort]
known-first-party = ["src", "scripts"]
